+++ embedded-graphics/src/lib.rs
@@ -488,7 +488,7 @@
 #[cfg(feature = "nalgebra_support")]
 extern crate nalgebra;
 
-mod draw_target;
+pub mod draw_target;
 pub mod drawable;
 pub mod fonts;
 pub mod geometry;
@@ -499,5 +499,3 @@ pub mod prelude;
 pub mod primitives;
 pub mod style;
 pub mod transform;
-
-pub use draw_target::DrawTarget;