+++ embedded-graphics/src/prelude.rs
@@ -1,6 +1,7 @@
 //! Prelude
 
 pub use crate::{
+    draw_target::DrawTarget,
     drawable::{Drawable, Pixel},
     fonts::Font,
     geometry::{Dimensions, Point, Size},
@@ -8,5 +9,4 @@ pub use crate::{
     pixelcolor::{raw::RawData, GrayColor, IntoStorage, PixelColor, RgbColor},
     primitives::Primitive,
     transform::Transform,
-    DrawTarget,
 };